requirements.txt
asgiref==3.9.2
Django==5.2.6
django-cors-headers==4.9.0
django-jazzmin==3.0.1
djangorestframework==3.16.1
drf-yasg==1.21.11
inflection==0.5.1
packaging==25.0
pillow==11.3.0
pytz==2025.2
PyYAML==6.0.3
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
